change pstate 16 to pstate 0
The llama.cpp patch tries to set the GPUs to high performance mode when beginning inference. However, it currently has the pstate values backwards, and is using 16 (the lowest) as if it were the highest. This actually works fine on my machine, at least, since `nvidia-pstate -ps 16` successfully forces pstate to... 0. For some reason. But might as well specify the right value in case that gets fixed.